[Dario] Included addNew property linked to addShadowUserOnLogin property for backward compatibility. Change formatting in SamlLoginIT.

+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/integration/feature/SamlLoginIT.java
@@ -473,6 +473,7 @@ public class SamlLoginIT {
         if (null == userId) {
             return;
         }
+        if (userId == null) return;
 
         IntegrationTestUtils.deleteUser(zoneAdminToken, baseUrl, userId);
     }
